+++ src/fonts.test.ts
@@ -1,59 +1,7 @@
 import { describe, expect, it } from 'vitest';
 
 import type { FontWeight } from './api/text.ts';
-import { readFonts, weightToNumber } from './fonts.ts';
-
-describe('readFonts', () => {
-  it('returns fonts array', () => {
-    const fontsDef = {
-      Test: [
-        { data: mkData('Test_Sans_Normal') },
-        { data: mkData('Test_Sans_Italic'), italic: true },
-        { data: mkData('Test_Sans_Bold'), bold: true },
-        { data: mkData('Test_Sans_BoldItalic'), italic: true, bold: true },
-      ],
-      Other: [{ data: mkData('Other_Normal') }],
-    };
-
-    const fonts = readFonts(fontsDef);
-
-    expect(fonts).toEqual([
-      { family: 'Test', style: 'normal', weight: 400, data: mkData('Test_Sans_Normal') },
-      { family: 'Test', style: 'italic', weight: 400, data: mkData('Test_Sans_Italic') },
-      { family: 'Test', style: 'normal', weight: 700, data: mkData('Test_Sans_Bold') },
-      { family: 'Test', style: 'italic', weight: 700, data: mkData('Test_Sans_BoldItalic') },
-      { family: 'Other', style: 'normal', weight: 400, data: mkData('Other_Normal') },
-    ]);
-  });
-
-  it('throws on missing input', () => {
-    expect(() => readFonts(undefined)).toThrow(new TypeError('Expected object, got: undefined'));
-  });
-
-  it('throws on invalid type', () => {
-    expect(() => readFonts(23)).toThrow(new TypeError('Expected object, got: 23'));
-  });
-
-  it('throws on invalid italic value', () => {
-    const fn = () => readFonts({ Test: [{ data: 'data', italic: 23 }] });
-
-    expect(fn).toThrow(
-      new TypeError('Invalid value for "Test/0/italic": Expected boolean, got: 23'),
-    );
-  });
-
-  it('throws on invalid bold value', () => {
-    const fn = () => readFonts({ Test: [{ data: 'data', bold: 23 }] });
-
-    expect(fn).toThrow(new TypeError('Invalid value for "Test/0/bold": Expected boolean, got: 23'));
-  });
-
-  it('throws on missing data', () => {
-    const fn = () => readFonts({ Test: [{ italic: true }] });
-
-    expect(fn).toThrow(new TypeError('Invalid value for "Test/0": Missing value for "data"'));
-  });
-});
+import { weightToNumber } from './fonts.ts';
 
 describe('weightToNumber', () => {
   it('supports keywords `normal` and `bold`', () => {
@@ -79,7 +27,3 @@ describe('weightToNumber', () => {
     expect(() => weightToNumber(0.1)).toThrow(new Error('Invalid font weight: 0.1'));
   });
 });
-
-function mkData(value: string) {
-  return new Uint8Array(value.split('').map((c) => c.charCodeAt(0)));
-}